+++ packages/orq-rc/RELEASES.md
@@ -4039,3 +4039,13 @@ Based on:
 - [typescript v3.14.0-rc.3] packages/orq-rc
 ### Releases
 - [NPM v3.14.0-rc.3] https://www.npmjs.com/package/@orq-ai/node/v/3.14.0-rc.3 - packages/orq-rc
+
+## 2025-10-02 14:56:32
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.633.2 (2.720.5) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.14.0-rc.4] packages/orq-rc
+### Releases
+- [NPM v3.14.0-rc.4] https://www.npmjs.com/package/@orq-ai/node/v/3.14.0-rc.4 - packages/orq-rc